requirements.txt
mcp>=1.1.0
googlemaps>=4.10.0
python-dotenv>=1.0.0